Modification to ntop's parameters to allow specification of the syslog facility via a --use-syslog=facility parameter.
Note that unlike the usual short/long mapping, -L and --use-syslog
are not exactly equivalent - see the comments in the code and
documentation file.

Note that to use the parameter it must be coded as:

     --use-syslog=facility

If you code:

     --use-syslog facility

the facility is ignored and it's treated just like -L (which uses a default
facility)

-----Burton (ref BMS0060)

PS: I was sure I had checked this in a while back, but can't find any
records... so here it is.


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@883 a02cd0c1-8e76-42e1-a119-56f9641475e2