Although the documentation would have led you to believe otherwise, the -w and -W parameters ignored any address they were given when binding to the sockets.  Thus  -w 192.168.0.1:3000 and -w 3000 behaved the same.  This caused unexpected behavior when multiple addresses were defined for the same interface, or multiple interfaces existed in the machine.
The code which used INADDR_ANY has been replaced to use the given value,
if it can be translated to an interface, with the log message:

Unable to convert address '%s'... Not binding to a particular interface!

A bogus value (or none at all) reverts to the previous, expected (?) behavior.

Under MinGW, which apparently doesn't support this facility, ntop continues
to use INADDR_ANY.

Finally, for completeness, the address was added to the "waiting for" message
text for the SSL connection, just like it was already present for non-SSL.

-----Burton (ref BMS-0077)

Thanks to Razvan Cosma [razvan.cosma@catv.telemach.ro] for also testing
this patch.


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@900 a02cd0c1-8e76-42e1-a119-56f9641475e2