I've been having problems w/ the ntop web server hanging when accessed via ssl (https://) from Netscape 6.2.2 (Win2K) (and others).
I put a fix into ntop to create a "watchdog".  It waits for 3 seconds and
then if the SSL_accept call (openSSL) hasn't finished, it aborts it.  This
leaves the user with nothing on their web browser, but at least ntop's web
server continues on.  I don't know of a way to send something back to the
user, because it's the browser-server handshake that's hung.  It looks - to
the user - like a failed connection.

The problem is that ntop's web server is single threaded until we determine
that the request is simply one that will be reading data.  At that point we
fork to generate the page.  But the basic "accept a request" code is single
threaded.  This happens all but instantaneously and hasn't been a problem
previously.

The problem boils down to a hang in a call, SSL_accept() that doesn't have
a timeout parameter.

Because the code is invasive, I've built it (like the SIGPIPE stuff) so you
can turn it on at ./configure time:

  --enable-sslwatchdog    Watchdog for ssl hangups (Netscape 6.2.2)
[default=disabled]

or via a command line option:

  --ssl-watchdog          Use ssl watchdog (NS6 problem)

With the "fix", ntop's web server hangs for at most 3 seconds, then
continues on.  The user gets nada - and I don't know a way to send them
anything, because we haven't retrieved the request yet nor done the
handshake (so there isn't a TCP connection!) Look for this in the log:

Jul  3 17:20:57 tigger ntop[20240]: SSLWDERROR: Watchdog timer has expired.
Aborting request, but ntop processing continues!

It only affects https:// requests and I've coded the watchdog so it doesn't
activate unless we have openSSL and either the compile or runtime parameter
set.  If you don't get https:// requests, it's just another idle thread.

Note that this version may not be final - Luca wants to see and test the
code, and it hasn't been tested under anything but Linux.  Solaris needs to
be tested - Win32 shouldn't be an issue (no threads)...

-----Burton  (Ref BMS-0101)


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@1015 a02cd0c1-8e76-42e1-a119-56f9641475e2