(Linux fix)
    <sad-face>
      Fix a silly problem with the gdchart handling (I had forgotten to add -L.../zlib-1.1.4
      and -lz to the variable set when using the ntop static build for zlib)
    </sad-face>

      Also fix a silly construct (not ours) in the libpng makefiles (lots of them)
      where the location of zlib was hardcoded as ../zlib

      It's silly because the versions of zlib usually untar into zlib-1.1.4 or so.

What makes these interesting is that they ONLY surface if you don't have zlib
installed on the system and depend on the static build that's part of ntop's
buildAll.sh command.  For 99.9% of users (obviously), they have a libz.a or
libz.so in /usr/lib or some other system place and never hit this...

I only found it because I had uninstalled the shared libz.so on my UML build
machine because only rpms for 1.1.3 were available from RedHat.

ser-en-dip-i-ty n. pl. ser-en-dip-i-ties
   The faculty of making fortunate discoveries by accident.
   The fact or occurrence of such discoveries.
   An instance of making such a discovery.

-----Burton


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@1184 a02cd0c1-8e76-42e1-a119-56f9641475e2