This enhancement adds a new tab, "Total" to the ntop menu bar.  The "Total" tab mirrors "Sent" and "Rcvd", but presents the total traffic to/from a host.  Reports are labeled "Network Traffic: Total Data (Sent+Received)"
As part of this:

* new counters were introduced for ntop run-to-date byte and packet
counts.  These are used in the calculation of average values -- preventing the
absurd values for "average" throughput when ntop is run with persistent data.

* ntop.h constants were added and used in all the case constructs.  These mirror
the existing STR_SORT_DATA_SENT_PROTOS with a SORT_DATA_RECEIVED_PROTOS #define.

* A report footer was added to some of the reports, explaining how the values are
calculated and/or what they mean.

This addresses a request made on SourceForge:
     #488948 extra button "Data Sent+Rcvd" * 2001-12-04 08:40 nobody bluestorm

-----Burton (ref 130)


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@1200 a02cd0c1-8e76-42e1-a119-56f9641475e2