If you ran ntop with -S and -M and then ran ntop with only -S, it was possible for the persistent value of actualReportDeviceId to specify a value > 0 and so the reporting would be pointed at a device[] entry that was not being populated with data.
This forces actualReportDeviceId to zero if we're merging interfaces.

Similarly, if you have a stored actualReportDeviceId greater than the
actual number of devices (say you ran ntop with -i a,b,c and switched
reporting to c, then reran with only -i a,b), the value could be out
of range.

That case was already fixed, but this patch adds a note to the log:

Note: stored actualReportDeviceId(%d) > numDevices(#). Probably leftover, reset.

-----Burton (ref 141)


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@1225 a02cd0c1-8e76-42e1-a119-56f9641475e2