This has torqued me since I first started with ntop.
Fix the generated html code for the following "charts" on
the "Info about host xxxxx" page:

  Data Sent Stats
  IP vs. Non-IP Sent
  Data Rcvd Stats
  IP vs. Non-IP Rcvd
  Sent vs. Rcvd Pkts
  Sent vs. Rcvd Data

And some on the "IP Protos" "Distribution" page.

For the old version, the browser would allocate too much space
for small values (enough to hold the text).

The new version allocates separate cells for the legends and
divides the space up according to the actual distribution.

It also rounds the values being displayed vs. truncating them.


NOTE: The generated code will not display properly under Opera 6.
      I was able to cut it down to a single page of trivial html
      (with lots of white space) that validates according to W3C
      and still renders improperly under Opera 6.

      I have reported the bug and been told it's fixed in the
      Opera 7 beta.

      This HTML has been tested and works fine on NS 4.61+,
      IE 5.5+, Mozilla 0.99+, Galeon and Konqueror.

NOTE: If you don't like it, there is an #if 0/#if 1 switch in
      reportUtils.c, but that's going to get ripped out after
      a week or so, as long as the code works.

NOTE: Yes, there is a SMALL increase in the "size" of the page.
      Since it's already big and you have to scroll anyway, s'be'it.

-----Burton (ref 133)


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@1243 a02cd0c1-8e76-42e1-a119-56f9641475e2