Updates to new configure files for automake 1.7
1. Removed rule creating aclocal.m4 -- it's created in makefile.in automatically.
2. Moved contents of acinclude.m4 -> acinclude.m4.ntop (these are the unique ntop
   macros)
3. Added Makefile rule to rebuild acinclude.m4 from acinclude.m4.ntop + libtool.m4.in
4. Changed CFLAGS and LDFLAGS to AM_CFLAGS and AM_LDFLAGS, since automake 1.7 now does
   CFLAGS and LDFLAGS automagically.

Note that the lines added to acinclude.m4.ntop beyond what used to be acinclude.m4
are a bug fix from the automake cvs -- and will probably be part of automake 1.7.1.
I *think* we'll need to keep them in ntop's file until there is no chance of ever
running into automake 1.7...

With these changes, the scripts work - for Luca and I - under the various combinations
of automake/autoconf, that is:

autoconf 2.53/automake 1.6.3 (i.e. the RedHat 7.x and 8.x defaults)
autoconf 2.54/automake 1.7   (build from scratch)

plus other combinations on other platforms.

The scripts will NOT work with autoconf 2.13

libtool 1.4+ is required.

-----Burton


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@1284 a02cd0c1-8e76-42e1-a119-56f9641475e2