(Note to Luca - I've only tested under Linux, so if it doesn't work,  well, just back it out (it's self contained).  However, it should  work - this is generic code from the GNU C documentation)
This change implements a "watchdog" on the gdchart creation of pie charts.

If you have a compile/runtime version mis-match, there is an abort() call
in libpng, which kills the web browser thread. Bad...

This change uses the fork() call to let the gdchart creation run in a
separate thread.  It uses the parent thread as a watchdog to monitor
a status variables to detect failure.

If the generation of the chart fails, a billboard graphic, pie-error.png
is sent to the user instead.  The specific failure is logged for
the admin to use.

-----Burton (ref 149)


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@1298 a02cd0c1-8e76-42e1-a119-56f9641475e2