Update the code to properly push out the generated graph files.
Make the new code the default.

It's been tested under FreeBSD and Linux.  Pls report problems to ntop-dev.

If you need to disable it, there is a switch in ntop.h, just uncomment:

/* #define DISABLE_GDC_WATCHDOG */


If the libpng conflict exists on your system, instead of crashing the ntop
web server, this code detects it and will push out an "error" graphic.

-----Burton (ref 149, part2)


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@1302 a02cd0c1-8e76-42e1-a119-56f9641475e2