Using the code from GNU libiberty, enable getopt_long() for those systems which don't have it in their native gcc compiler.
Basically, if the HAVE_GETOPT_LONG flag is NOT set, then code in util.c
- taken from libiberty (which is a library dedicated to providing glibc
library functions on systems which don't already have them) - provides it.

So it should use the system routine if available, or if unavailable, ours.

The only known situation for this was FreeBSD 4.6.2, and this has been tested
there (as well as under Linux to make sure it doesn't break existing cases).

(NB: I'm not sure about FreeBSD 4.7)

-----Burton (ref 156)


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@1312 a02cd0c1-8e76-42e1-a119-56f9641475e2