Automate the handling of the automake 1.7.1 fix.
If ./configure has made the necessary changes for 1.7.1, it creates
a sentinal file and so after it creates the files, a standard ./configure
(with whatever parameters are necessary) works.

-----Burton (ref 162)

NB: The diff looks like a lot of lines, but that's because I had to shove
them over for indentation inside an if.


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@1329 a02cd0c1-8e76-42e1-a119-56f9641475e2