Fix a deadlock in ntop.
This was most commonly seen as the web server apparently locking up or
failing to respond to queries.  A detailed look via gdb showed almost all
the threads suspended waiting for different mutexes.

This was very common in RedHat 8/9, less so in other OSes.

Also corrected the 'title' of the accessMutex() call in http.c

-----Burton


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@1909 a02cd0c1-8e76-42e1-a119-56f9641475e2