Fix a deadlock during shutdown.
Luca added a lock of the purge mutex during page creation so we don't
free the data in the middle of reporting on it.  However, there's
also a lock of that mutex in the cleanup routine (to prevent purge
during cleanup).  That lock of a self-locked mutex was deadlocking
shutdown (on those systems that don't just ignore lock of self-locked).

Added a release() before calling shutdown() - this allows any pending purge
to terminate, then does the cleanup.

Also added a log message of the request and messages on either side of the
accessMutex call during shutdown - so the user won't think it hung:

   Shutdown request has been received [MSGID0538544]
...
   CLEANUP: Locking purge mutex (may block for a little while) [MSGID8844426]
   CLEANUP: Locked purge mutex, continuing shutdown [MSGID8814988]

If for some inexplicable reason shutdownNtop() returns to it's caller,
that's logged and reported as part of the web page sent back to the user.

   Shutdown request FAILED [MSGIDtbd]

-----Burton (ref 341)


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@1971 a02cd0c1-8e76-42e1-a119-56f9641475e2