Add -j | --create-other-packets option.
This creates an output file of those pesky 'other' traffic packets so you
can analyze them to figure out what they are.

-----Burton

This was contributed by Loic [50891@snr.fr] and I had previously indicated
I was going to hold it post 2.3.  But enough time and other changes have
been put in that the reason for holding it (code freeze) isn't valid.
Sorry for sitting on this so long...


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@2152 a02cd0c1-8e76-42e1-a119-56f9641475e2