Fixed buffer length problem when IP ports are displayed in the HTML pages
Improved the -g option: ntop now partially keeps track of remote hosts. Basically
it reports the hosts contacted but those hosts are not really present in ntop. The
outcome is that the page report is more dense but if you click on a remote host
you won't be able to display any further info (as you used -g). This is useful
for those who monitor a large network and want to track local hosts but have
a clue what remote hosts are contacted.

Added a limit of 4096 address entries queued waiting to be resolved. This
prevents that a DoS app uses ntop to fill-up all the available disk space.


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@2165 a02cd0c1-8e76-42e1-a119-56f9641475e2