BMS0456
Fixup handling of the time stamp on the .db files vs. the .tgz files.
It turns out that we need to compare against the latest of the ctime and
mtime (create and modified) as the cp done by make install updates the
ctime but not the mtime (which was all we were checking).

Thus if you updated the file and did a make install, ntop didn't reload
the database.

-----Burton


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@2345 a02cd0c1-8e76-42e1-a119-56f9641475e2