BMS0475
Fix (?) the dns 'problem' some users are having.  There was confusion
in the code as to which buffer was key and which was data. Renamed them
to make the relationship explicit (and fix the case where the wrong
one was being used).  Also made sure that the buffer is long enough
for a max IPv6 address.

Note that it is strongly recommended that you delete the existing file,
dnsCache.db whenever installing a newer version of ntop (See BMS0476 for
that)

-----Burton


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@2398 a02cd0c1-8e76-42e1-a119-56f9641475e2