mVLAN:
This patch allows ntop to recognize that it is seeing the same host on
multiple VLANs (host is matched by the combination of IP + MAC) on the
same monitored interface.

Most users will never have seen the problem - it's only if your network
is as screwy as mine is ... and while it's not WRONG, it's definitely odd
to ntop.  See the discussion in docs/FAQ.

Why bother? It's one of the causes for the illegal rrd_update problem. Two
HostTraffic entries for the "same" host will cause multiple updates of the
rrd file to be attempted and cause LOTS of error messages.

If found, a single error message is issued:

     **ERROR** mVLAN: Host (identical IP/MAC) found on multiple VLANs
     mVLAN: ntop continues but will consolidate and thus probably overcount this traffic
     mVLAN: Up to 10 examples will be printed

And (as discovered) up to 10 specific cases will be identified:

     mVLAN: Host 192.168.aaa.bbb (00:12:17:xx:yy:zz) VLANs 3 and 4

-----Burton (ref 495)


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@2716 a02cd0c1-8e76-42e1-a119-56f9641475e2