Add packet delay timing report to info.html.
It's a moving average of last 1K (adjustable) packets, showing time
between libpcap time stamp (receipt) and start of processPacket().

The calculation in the note gives a VERY ROUGH approximation of the
rate of packet arrival ntop can handle.

In globals-defines.h, #undef MAX_ARRIVAL_BUFFER to disable this feature!

-----Burton (ref 571)


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntop@2763 a02cd0c1-8e76-42e1-a119-56f9641475e2